+++ wifipumpkin3/core/servers/mitm/responder3.py
@@ -66,7 +66,9 @@ class Responder3(MitmMode):
     @property
     def CMD_ARRAY(self):
         iface = self.conf.get("accesspoint", "interface")
-        config_responder3_path = C.user_config_dir + self.conf.get("mitm_modules", "responder3_config")
+        config_responder3_path = C.user_config_dir + self.conf.get(
+            "mitm_modules", "responder3_config"
+        )
         self._cmd_array = ["-I", iface, "-4", "-p", config_responder3_path]
         return self._cmd_array